% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{area,STGrid-method}
\alias{area,STGrid-method}
\title{Get the area of an STGrid object.}
\usage{
\S4method{area}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Area is computed as (x_max - x_min)*(y_max - y_min).
}
\examples{
example_dataset()
area(Xenium_Mouse_Brain_Coronal_7g)
}
