% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_mat,STGrid-method}
\alias{bin_mat,STGrid-method}
\title{The binned matrix stored in a STGrid object.}
\usage{
\S4method{bin_mat}{STGrid}(
  object,
  as_factor = FALSE,
  gene_list = character(),
  melt_tab = FALSE,
  del_bin = FALSE,
  all_genes = FALSE
)
}
\arguments{
\item{object}{The STGrid object}

\item{as_factor}{Should bin_x and bin_y columns be returned as ordered factor ?}

\item{gene_list}{Whether to subset to some genes.}

\item{melt_tab}{Whether to melt.}

\item{del_bin}{Whether to delete the bin_x, bin_y columns.}

\item{all_genes}{Whether to delete the 'all_genes' columns.}
}
\description{
The binned matrix stored in a STGrid object.
}
