% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_mat}
\alias{bin_mat}
\title{The binned matrix stored in a STGrid object}
\usage{
bin_mat(object, as_factor = FALSE, gene_list = character(), melt = FALSE)
}
\arguments{
\item{object}{The STGrid object}

\item{as_factor}{Should bin_x and bin_y columns be returned as ordered factor ?}

\item{gene_list}{Whether to subset to some genes.}

\item{melt}{Whether to melt.}
}
\description{
The binned matrix stored in a STGrid object
}
\keyword{internal}
