% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_size,STGrid-method}
\alias{bin_size,STGrid-method}
\title{The size of the bins stored in an STGrid object}
\usage{
\S4method{bin_size}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The size of the bins stored in an STGrid object
}
\examples{
example_dataset()
bin_size(Xenium_Mouse_Brain_Coronal_7g)
}
