% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_size}
\alias{bin_size}
\title{The size of the bins stored in an STGrid object}
\usage{
bin_size(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The size of the bins stored in an STGrid object
}
\examples{
example_dataset()
bin_size(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
