% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_this_matrix}
\alias{bin_this_matrix}
\title{Create a 2D binned grid (rasterization) from molecule coordinates}
\usage{
bin_this_matrix(coord = NULL, bin_size = 25, control = NULL)
}
\arguments{
\item{coord}{A data frame containing molecule coordinates with columns 'x', 'y', and 'gene'. Defaults to NULL.}

\item{bin_size}{An integer specifying the size of each bin. Defaults to 25.}
}
\value{
A list containing the binned spatial matrix, updated molecule coordinates, and information about the binning process.
}
\description{
This function takes molecule coordinates as input (with columns named 'x', 'y', 'gene') and creates a 2D
binned grid by dividing the x and y axes into bins. It returns the counts of the number of molecules in
each bin along with additional information about the binning process.
}
\examples{
# Create a binned grid from molecule coordinates
bin_mat(coord = my_coord, bin_size = 25)
}
\keyword{internal}
