% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_x,STGrid-method}
\alias{bin_x,STGrid-method}
\title{X bins of a STGrid object.
row_names}
\usage{
\S4method{bin_x}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
X bins of a STGrid object.
}
