% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_y,STGrid-method}
\alias{bin_y,STGrid-method}
\title{Y bins of a STGrid object.
row_names}
\usage{
\S4method{bin_y}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Y bins of a STGrid object.
}
