% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_y}
\alias{bin_y}
\title{Y bins of a STGrid object.}
\usage{
bin_y(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Y bins of a STGrid object.
}
\examples{
example_dataset()
bin_y(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
