% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{$<-,STGrid-method}
\alias{$<-,STGrid-method}
\title{Replacing operator "$" of a STGrid object. Extract metadata.}
\usage{
\S4method{$}{STGrid}(x, name) <- value
}
\arguments{
\item{value}{The replacement value.}

\item{i}{indices specifying metadata to extract.}

\item{exact}{See ?"$".}
}
\description{
Subsetting operator "$" of a STGrid object. Extract metadata.
}
\examples{
example_dataset()
Xenium_Mouse_Brain_Coronal_7g$count_sum <- Xenium_Mouse_Brain_Coronal_7g$count_sum
}
\keyword{internal}
