% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{$,STGrid-method}
\alias{$,STGrid-method}
\title{Subsetting operator "$" of a STGrid object. Extract metadata.}
\usage{
\S4method{$}{STGrid}(x, name)
}
\arguments{
\item{name}{The metadata to extract.}

\item{i}{indices specifying metadata to extract.}
}
\description{
Subsetting operator "$" of a STGrid object. Extract metadata.
}
\examples{
example_dataset()
head(Xenium_Mouse_Brain_Coronal_7g$count_sums)
head(Xenium_Mouse_Brain_Coronal_7g[[1]])
}
\keyword{internal}
