% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{check_features_exist,STGrid-method}
\alias{check_features_exist,STGrid-method}
\title{Check all specified features exist in  an STGrid object.}
\usage{
\S4method{check_features_exist}{STGrid}(object, feat_list, no_meta = FALSE)
}
\arguments{
\item{object}{The STGrid object}

\item{feat_list}{The list of features to be tested.}

\item{no_meta}{Whether meta features should not be considered. Default is FALSE.}
}
\description{
Check all specified features exist in  an STGrid object.
}
\examples{
example_dataset()
check_features_exist(Xenium_Mouse_Brain_Coronal_7g, "Ano1")
check_features_exist(Xenium_Mouse_Brain_Coronal_7g, "foo")
}
