% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{check_st_list}
\alias{check_st_list}
\title{Check Validity of STGrid List and Feature List}
\usage{
check_st_list(st_list, feat_list = NULL)
}
\arguments{
\item{st_list}{A list of objects to be validated, where each object should be
of class 'STGrid'.}

\item{feat_list}{An optional vector of feature names to be checked across
all STGrid objects in 'st_list'. If 'NULL', only the STGrid object
validation is performed.}
}
\value{
Invisible NULL. The function is called for its side effect of
stopping execution with an error message if any validation fails.
}
\description{
Validates a list of STGrid objects and an optional list of features.
Ensures that the list contains at least one STGrid object, all objects in the
list are of type STGrid, and all specified features exist in the STGrid objects.
}
\examples{
example_dataset()
check_st_list(list(Xenium_Mouse_Brain_Coronal_7g, Xenium_Mouse_Brain_Coronal_7g))
}
\keyword{internal}
