% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_bar_plot,STCompR-method}
\alias{cmp_bar_plot,STCompR-method}
\title{Create a barplot to show counts for selected features.}
\usage{
\S4method{cmp_bar_plot}{STCompR}(
  object,
  features = head(feat_names(object)),
  normalized = FALSE,
  transform = c("None", "log2", "log10", "log"),
  colors = c("#3074BB", "#BE5B52")
)
}
\arguments{
\item{object}{A STGrid object.}

\item{features}{The list of features (NULL for all of them).}

\item{normalized}{Whether counts should be normalized.}

\item{transform}{Whether the count should be transformed (the pseudo count defined for the object is added).}

\item{colors}{A set of colors.}
}
\description{
Create a barplot to show counts for selected features.
}
