% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_boxplot,STCompR-method}
\alias{cmp_boxplot,STCompR-method}
\title{Create a boxplot/jitter plot to show molecule counts distribution.}
\usage{
\S4method{cmp_boxplot}{STCompR}(
  object,
  normalized = TRUE,
  transform = c("None", "log2", "log10", "log"),
  colors = c("#3074BB", "#BE5B52"),
  ...
)
}
\arguments{
\item{object}{A STCompR object.}

\item{normalized}{A logical value indicating whether to use normalized counts.}

\item{transform}{A character string specifying the transformation to be applied to molecule counts.
Options include "None", "log2", "log10", and "log".}

\item{colors}{A vector of colors for the boxplot/jitter plot.}

\item{...}{Additional arguments to be passed to the underlying ggpol::geom_boxjitter function.}
}
\value{
A ggplot object displaying the molecule counts distribution.
}
\description{
Create a boxplot/jitter plot to show molecule counts distribution.
}
\examples{
# Example usage:
st_grid_1 <- create_STGrid(...)
st_grid_2 <- create_STGrid(...)
st_compr_result <- stcompr(st_grid_1, st_grid_2, name_1 = "Condition1", name_2 = "Condition2")
cmp_boxplot(st_compr_result, normalized = TRUE, transform = "log2", colors = c("blue", "red"))

}
