% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_boxplot}
\alias{cmp_boxplot}
\title{Create a boxplot/jitter plot to show molecule counts distribution.}
\usage{
cmp_boxplot(
  object_1,
  object_2,
  genes = head(gene_names(object_1)),
  name_1 = "Condition_1",
  name_2 = "Condition_2",
  colors = c("#4DBBD5B2", "#F39B7FB2"),
  ...
)
}
\arguments{
\item{object_1}{A STGrid object.}

\item{object_2}{A STGrid object.}
}
\description{
Create a boxplot/jitter plot to show molecule counts distribution.
}
\keyword{internal}
