% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_dimensions}
\alias{cmp_dimensions}
\title{Compare Areas of Spatial Transcriptomics (STGrid) Objects}
\usage{
cmp_dimensions(
  ...,
  type = c("area", "x_size", "y_size"),
  fill_color = NULL,
  border_color = "black"
)
}
\arguments{
\item{...}{One or more \code{STGrid} objects or a list of \code{STGrid} objects to be compared.}

\item{type}{What should be computed: area, x size or y size.}

\item{fill_color}{A vector specifying the fill colors for the bars, corresponding to the samples. Defaults to \code{NULL}, in which case default ggplot colors are used.}

\item{border_color}{A color for the border of the bars. Defaults to \code{"black"}.}
}
\value{
A ggplot2 object representing a bar plot of the areas for the provided STGrid objects.
}
\description{
This function displays the areas of multiple STGrid objects using a bar plots.
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
x_bins <-  bin_x(xen)[181:nbin_x(xen)]
y_bins <-  bin_y(xen)[101:nbin_y(xen)]
xen_1 <- xen[x_bins, y_bins]
x_bins <-  bin_x(xen)[61:101]
y_bins <-  bin_y(xen)[101:nbin_y(xen)]
xen_2 <- xen[x_bins, y_bins]
cmp_dimensions(xen_1, xen_2)
}
