% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_images}
\alias{cmp_images}
\title{Compare Images}
\usage{
cmp_images(
  ...,
  feat_list = NULL,
  names = NULL,
  colors = viridis::inferno(10),
  saturation = 1,
  coord_fixed = TRUE,
  scale = TRUE,
  logb = 10,
  pseudo_count = 1,
  condition_vs_feat = TRUE
)
}
\arguments{
\item{...}{STGrid objects to be compared.}

\item{feat_list}{A list of features for which the spatial images will be created.}

\item{names}{Names for the conditions or experiments. Defaults to NULL.}

\item{colors}{The colors to use for gradient fill in the spatial images. Defaults to viridis::inferno(10).}

\item{saturation}{The ceiling level for the feature expression values. Defaults to 1 (no ceiling).}

\item{coord_fixed}{Logical value indicating whether to keep the aspect ratio fixed. Defaults to TRUE.}

\item{scale}{Logical value indicating whether to scale the feature expression values. Defaults to TRUE.}

\item{logb}{The basis for the log transformation. Default to 10. If NULL, no log transformation.}

\item{pseudo_count}{A value for the pseudo count used for log transformation (default to 1).}

\item{condition_vs_feat}{Logical indicating whether to facet by condition vs. feature (TRUE) or by feature vs. condition (FALSE). Defaults to TRUE.}
}
\description{
This function compares and visualizes multiple spatial images obtained from STGrid objects. It displays a color-coded representation of the feature (e.g., molecules) density observed in spatial transcriptomics experiments for different conditions.
}
