% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_mol_counts,STCompR-method}
\alias{cmp_mol_counts,STCompR-method}
\title{Create a barplot to show molecule counts of selected genes}
\usage{
\S4method{cmp_mol_counts}{STCompR}(
  object,
  genes = head(gene_names(object_1)),
  name_1 = "Condition_1",
  name_2 = "Condition_2",
  colors = c("#4DBBD5B2", "#F39B7FB2")
)
}
\arguments{
\item{object_1}{A STGrid object.}

\item{object_2}{A STGrid object.}
}
\description{
Create a barplot to show molecule counts of selected genes
}
