% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_mol_counts}
\alias{cmp_mol_counts}
\title{Create a barplot to show molecule counts of selected genes}
\usage{
cmp_mol_counts(
  object,
  genes = head(gene_names(object_1)),
  name_1 = "Condition_1",
  name_2 = "Condition_2",
  colors = c("#4DBBD5B2", "#F39B7FB2")
)
}
\arguments{
\item{object}{The STGrid object.}
}
\description{
Create a barplot to show molecule counts of selected genes
}
\keyword{internal}
