% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{col_names,STGrid-method}
\alias{col_names,STGrid-method}
\title{Column names of a STGrid object.}
\usage{
\S4method{col_names}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The column names of a STGrid object.
}
