% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{col_names}
\alias{col_names}
\title{Column names of a STGrid object.}
\usage{
col_names(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The column names of a STGrid object.
}
\keyword{internal}
