% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{compute_k_ripley,STGrid-method}
\alias{compute_k_ripley,STGrid-method}
\title{Estimate Ripley's reduced second moment function for each gene in a spatial grid.}
\usage{
\S4method{compute_k_ripley}{STGrid}(
  object,
  rmax = 200,
  nlarge = 1e+06,
  var.approx = FALSE,
  ratio = FALSE
)
}
\arguments{
\item{object}{An object of class "STGrid".}

\item{rmax}{Maximum desired value of the argument r.}

\item{nlarge}{Efficiency threshold. If the number of points exceeds nlarge, then only the border correction will be computed (by default), using a fast algorithm.}

\item{var.approx}{Logical. If TRUE, the approximate variance of K(r) under CSR will also be computed.}

\item{ratio}{Logical. If TRUE, the numerator and denominator of each edge-corrected estimate will also be saved, for use in analyzing replicated point patterns.}
}
\value{
An updated object of class "STGrid" with the Ripley's K function estimates stored in the slot 'ripley_k_function'.
}
\description{
This method calculates Ripley's reduced second moment function, K(r), for each gene in a spatial grid.
}
\examples{
# Example usage:
# grid_object <- load_spatial(...)
# grid_object <- compute_k_ripley(grid_object)

}
