% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{compute_module_score,STGrid-method}
\alias{compute_module_score,STGrid-method}
\title{From a list of feature (e.g gene) modules, compute a score (e.g. mean value) accross bins.}
\usage{
\S4method{compute_module_score}{STGrid}(object, modules = NULL, rename = FALSE, fun = "mean")
}
\arguments{
\item{object}{An \code{STGrid} object.}

\item{modules}{A list of modules.}

\item{rename}{Force the renaming of the modules to module_1, module_2...}

\item{fun}{The function to be applied (e.g. mean, median, sd...)}
}
\description{
From a list of feature (e.g gene) modules, compute a score (e.g. mean value) accross bins.
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
p <- hc_tree(xen, class_nb = 4, class_name = letters[1:4])
print(p)
# Get classes:
xen <- compute_module_score(xen, modules=p$tree_classes)
}
