% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{compute_module_score}
\alias{compute_module_score}
\title{From a list of feature (e.g gene) modules, compute a score (e.g. mean value) accross bins.}
\usage{
compute_module_score(object, modules = NULL, rename = FALSE, fun = "mean")
}
\arguments{
\item{object}{An \code{STGrid} object.}

\item{modules}{A list of modules.}

\item{rename}{Force the renaming of the modules to module_1, module_2...}

\item{fun}{The function to be applied (e.g. mean, median, sd...)}
}
\description{
From a list of feature (e.g gene) modules, compute a score (e.g. mean value) accross bins.
}
\keyword{internal}
