% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{compute_moran_index,STGrid-method}
\alias{compute_moran_index,STGrid-method}
\title{Compute Moran's Index.}
\usage{
\S4method{compute_moran_index}{STGrid}(
  object = NULL,
  queen_or_rook = c("queen", "rook"),
  scaling = TRUE
)
}
\arguments{
\item{queen_or_rook}{Method for neighborhood computation. Default to 'rook'. May also be queen.}

\item{scaling}{Whether to scale so that the sum of counts equals to 100 for each feature. Default TRUE.}
}
\description{
Compute Moran's Index.
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
xen <- compute_moran_index(re_bin(xen, bin_size = 400))
print(xen@moran_index)
}
