% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{connected_components}
\alias{connected_components}
\title{Given a Feature and an STGrid object, Label and Store Connected Components.}
\usage{
connected_components(
  object = NULL,
  feat_list = NULL,
  threshold = NULL,
  min_size = 9,
  neighborhood = 0,
  merge = TRUE,
  labels = NULL
)
}
\arguments{
\item{object}{An STGrid object containing the spatial expression data.}

\item{feat_list}{A character vector specifying the features of interest.}

\item{threshold}{A numeric value specifying the threshold for binarizing the matrix. If NULL, the matrix should already be binary.}

\item{min_size}{An integer specifying the minimum size of the connected components to retain. Default is 9.}

\item{neighborhood}{Extend the size of the neighborhood for each non zero value before component analysis.}

\item{merge}{Logical. If TRUE, label all connected components with the same label ("1").}

\item{labels}{A set of labels to be used for the connected components.}
}
\value{
The input STGrid object with additional fields for each feature's connected components.
}
\description{
This function identifies and labels connected components in a binary matrix representation of features of an STGrid object.
The connected components are determined using the specified threshold and a minimum size for the components.
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
xen <- connected_components(xen, feat_list = feat_names(xen)[1:5], threshold = 1, min_size = 9)
head(meta(xen))
spatial_image(xen, feat=meta_names(xen)[-1], ncol=2, as_factor=TRUE, colors = c("black", rainbow(60)))
}
\keyword{internal}
