% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{coord,STGrid-method}
\alias{coord,STGrid-method}
\title{The coordinates stored in a STGrid object}
\usage{
\S4method{coord}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The coordinates stored in a STGrid object
}
