% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{coord}
\alias{coord}
\title{The coordinates stored in a STGrid object}
\usage{
coord(object, as_factor = FALSE)
}
\arguments{
\item{object}{The STGrid object}

\item{as_factor}{Should bin_x and bin_y columns be returned as ordered factor ?}
}
\description{
The coordinates stored in a STGrid object
}
\examples{
example_dataset()
head(coord(Xenium_Mouse_Brain_Coronal_7g))
}
\keyword{internal}
