% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_hull,STGrid-method}
\alias{create_hull,STGrid-method}
\title{Create a Hull around a Feature component of an STGrid}
\usage{
\S4method{create_hull}{STGrid}(object = NULL, feature = NULL, color = "white", linewidth = 0.5)
}
\arguments{
\item{object}{An STGrid object containing the spatial expression data.}

\item{feature}{A character string specifying the meta feature of interest. Only one feature should be provided.}

\item{color}{A character string specifying the color of the hull lines. Default is "white".}

\item{linewidth}{A numeric value specifying the width of the hull lines. Default is 0.5.}
}
\value{
A list of ggplot2 geom_segment objects representing the convex hull. This elements can be added
to a ggplot object as produced by spatial_image().
}
\description{
The component of a binarized feature can be produced by the connected_components() function. This function creates a hull around
these components.
}
\details{
The function first checks if the specified meta feature exists in the STGrid object and whether it is numeric and integer.
It then generates a binary matrix representing the feature's presence, constructs horizontal and vertical segments
for the hull, and returns these segments as ggplot2 geom_segment objects.
}
\examples{
library(patchwork)
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
p1 <- spatial_image(xen, feat="Necab2")
xen <- connected_components(xen, feat_list = "Necab2", threshold = 5, min_size = 30)
head(meta(xen))
p2 <- spatial_image(xen, feat="Necab2_cpt",  as_factor=TRUE, colors = c("black", rainbow(9)))
hull <- create_hull(xen, feat="Necab2_cpt", color="red", linew=0.3)
p3 <- p2 + hull
p4 <- spatial_image(xen, feat="Nwd2") + hull
(p1 | p2 ) / (p3 | p4)
}
