% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{.DollarNames.STGrid}
\alias{.DollarNames.STGrid}
\title{This functions makes STGrid objects look like list and allow completion after $.}
\usage{
\method{.DollarNames}{STGrid}(x, pattern = "")
}
\arguments{
\item{x}{The STGtrid object.}

\item{pattern}{The pattern to be searched.}
}
\description{
This functions makes STGrid objects look like list and allow completion after $.
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
xen$bla <- "foo"
head(xen@meta)
}
\keyword{internal}
