% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_dataset}
\alias{example_dataset}
\title{Load/download a Seurat or ClusterSet example dataset.}
\usage{
example_dataset(
  dataset = c("10818896/files/Xenium_Mouse_Brain_Coronal_7g", "11210787/files/cmp_xen"),
  timeout = NULL,
  force = FALSE
)
}
\arguments{
\item{dataset}{The name of the dataset.}

\item{timeout}{Set the timout for download (options(timeout=timeout))}

\item{force}{Reload the dataset even if already in globalEnv.}
}
\value{
Load the dataset.
}
\description{
Load/download a Seurat or ClusterSet example dataset.
}
\examples{
example_dataset()
example_dataset("11210787/files/cmp_xen")
}
