% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{feat_names,STCompR-method}
\alias{feat_names,STCompR-method}
\title{The features stored in a STCompR object}
\usage{
\S4method{feat_names}{STCompR}(object)
}
\arguments{
\item{object}{The STCompR object}
}
\description{
The features stored in a STCompR object
}
\examples{
example_dataset("11284296/files/cmp_xen")
feat_names(cmp_xen)
}
