% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{feat_names,STGrid-method}
\alias{feat_names,STGrid-method}
\title{The features stored in an STGrid object}
\usage{
\S4method{feat_names}{STGrid}(object, del_control = FALSE)
}
\arguments{
\item{object}{The STGrid object.}

\item{del_control}{Whether to delete controls.}
}
\description{
The features stored in a STGrid object
}
