% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R, R/st_compr_class.R
\name{feat_names}
\alias{feat_names}
\title{The features stored in a STGrid object}
\usage{
feat_names(object, del_control = FALSE)
}
\arguments{
\item{object}{The STCompR object}

\item{del_control}{Whether to delete controls.}
}
\description{
The features stored in a STGrid object

The features stored in a STCompR object
}
\examples{
example_dataset()
feat_names(Xenium_Mouse_Brain_Coronal_7g)
example_dataset("11284296/files/cmp_xen")
feat_names(cmp_xen)
}
\keyword{internal}
