% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{feature_contrast}
\alias{feature_contrast}
\title{Gene Contrast Analysis for an STgrid object.}
\usage{
feature_contrast(
  object,
  feat_list_1 = NULL,
  feat_list_2 = NULL,
  low_color = "green",
  mid_color = "yellow",
  high_color = "red",
  na_color = "black",
  midpoint = 0,
  pseudo_count = 1,
  centered = TRUE,
  trim_ratio = 0.025
)
}
\arguments{
\item{object}{An STGrid object representing the spatial transcriptomics data.}

\item{feat_list_1}{A list of features (genes) for the first condition.}

\item{feat_list_2}{A list of features (genes) for the second condition.}

\item{low_color}{Colours for the low end of the gradient.}

\item{mid_color}{Colours for the midpoint of the gradient.}

\item{high_color}{Colours for the high end of the gradient.}

\item{na_color}{Bin with 0 value for both features are set to NA. The color for NA values.}

\item{midpoint}{The midpoint (in data value) of the diverging scale. Defaults to 0.}

\item{pseudo_count}{A small value added to avoid log transformation issues. Default is 1.}

\item{centered}{A boolean indicating whether to center the log2 ratio by subtracting the mean. Default is TRUE.}

\item{trim_ratio}{A ratio for trimming extreme log2 ratio values. Default is 0.025.}
}
\value{
A ggplot2 object representing the heatmap of log2 fold changes between the two feature sets.
}
\description{
This function performs a contrast analysis between two features (genes) using an STgrid object as input. It calculates the log2 fold change between the two features and visualizes the results as a rasterized image.
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
feature_contrast(xen, feat_list_1="Ano1", feat_list_2="Chat")
feature_contrast(xen, feat_list_1=c("Ano1", "Necab2"), feat_list_2=c("Chat", "Nrp2"))
}
\keyword{internal}
