% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{find_contiguous}
\alias{find_contiguous}
\title{Find contiguous cells using Queen criterion (internal).}
\usage{
find_contiguous(mat, threshold = 1)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{threshold}{A threshold for binarization.}
}
\value{
A list of matrices.
}
\description{
Find contiguous cells in a matrix using Queen criterion. The matrix is first binarized
(any value greater than zero is set to 1, any value lower than 1 is set to 0).
}
\examples{
set.seed(123)
m <- matrix(sample(0:2, replace=TRUE, size = 26*26, prob = c(0.98, 0.01, 0.01)), nc=26)
rownames(m) <- letters
colnames(m) <- LETTERS
image(m)
image(find_contiguous(m))
image(find_contiguous(m, ns=2)[[1]])
image(find_contiguous(m, threshold=2))
}
\keyword{internal}
