% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{gene_names,STCompR-method}
\alias{gene_names,STCompR-method}
\title{The genes stored in a STCompR object}
\usage{
\S4method{gene_names}{STCompR}(object)
}
\arguments{
\item{x}{The STCompR object}
}
\description{
The genes stored in a STCompR object
}
