% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{gene_names}
\alias{gene_names}
\title{The genes stored in a STGrid object}
\usage{
gene_names(object, all_genes = FALSE, del_control = FALSE)
}
\arguments{
\item{object}{The STGrid object}

\item{all_genes}{Whether "all_genes" should be included. Default to FALSE.}

\item{del_control}{Whether to delete controls.}
}
\description{
The genes stored in a STGrid object
}
\keyword{internal}
