% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{get_coord}
\alias{get_coord}
\title{The x/y coordinates of genes from a STGrid object.}
\usage{
get_coord(object, gene_list = character(), as.factor = TRUE)
}
\arguments{
\item{object}{The STGrid object}

\item{gene_list}{The list of genes}

\item{as.factor}{Whether the gene column should be returned as an ordered factor.}
}
\description{
Return the x/y coordinates of genes from a STGrid object
}
\keyword{internal}
