% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{get_gn_coord,STGrid-method}
\alias{get_gn_coord,STGrid-method}
\title{The x/y coordinates of genes from a STGrid object.}
\usage{
\S4method{get_gn_coord}{STGrid}(object, gene_list = character(), as.factor = TRUE)
}
\arguments{
\item{object}{The STGrid object}

\item{gene_list}{The list of genes}

\item{as.factor}{Whether the gene column should be returned as an ordered factor}
}
\description{
Return the x/y coordinates of genes from a STGrid object
}
