% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_verb_level}
\alias{get_verb_level}
\title{Get the current verbosity level.}
\usage{
get_verb_level()
}
\value{
A vector
}
\description{
This function get the verbosity level of the STarlight package which
controls the amount of information that is printed to the console by
the \code{\link{print_this_msg}} function.
}
\examples{
get_verb_level()
}
