% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{hc_tree,STGrid-method}
\alias{hc_tree,STGrid-method}
\title{Create a tree from an STGrid object}
\usage{
\S4method{hc_tree}{STGrid}(
  object,
  method = "complete",
  layout = "circular",
  dist_method = "pearson",
  branch_length = "none",
  class_nb = 1,
  class_name = "All",
  size = 2.25
)
}
\arguments{
\item{object}{The STGrid object}

\item{method}{The agglomeration method to be used. See stats::hclust().}

\item{layout}{The layout to be used See ggtree::ggtree().}

\item{dist_method}{The method for distance computation. See stats::cor().}

\item{branch_length}{A variable for scaling branch, if 'none' draw cladogram. See ggtree::ggtree().}

\item{class_nb}{An integer indicating the desired number of groups.}

\item{size}{The size of the labels.}
}
\description{
Create a tree from an STGrid object
}
\keyword{internal}
