% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{hc_tree}
\alias{hc_tree}
\title{Create a tree from an STGrid object}
\usage{
hc_tree(
  object,
  method = c("ward.D", "complete", "ward.D2", "single", "average", "mcquitty", "median",
    "centroid"),
  layout = c("circular", "rectangular", "slanted", "fan", "unrooted", "time-scale",
    "two-dimensional"),
  dist_method = "pearson",
  branch_length = "none",
  class_nb = 1,
  class_name = NULL,
  size = 2.25,
  colors = NULL
)
}
\arguments{
\item{object}{The STGrid object}

\item{method}{The agglomeration method to be used. See stats::hclust().}

\item{layout}{The layout to be used See ggtree::ggtree().}

\item{dist_method}{The method for distance computation. See stats::cor().}

\item{branch_length}{A variable for scaling branch, if 'none' draw cladogram. See ggtree::ggtree().}

\item{class_nb}{An integer indicating the desired number of groups.}

\item{size}{The size of the labels.}

\item{colors}{A set of colors for the classes.}
}
\description{
Create a tree from an STGrid object
}
\examples{
example_dataset()
p <- hc_tree(Xenium_Mouse_Brain_Coronal_7g)
print(p)
}
\keyword{internal}
