% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{heatmap_cmp}
\alias{heatmap_cmp}
\title{Generate a heatmap comparison for an 'STCompR' object}
\usage{
heatmap_cmp(
  object,
  what = c("changes", "changes_2", "c1", "c2"),
  hclust_method = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  dist_method = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  del_feat = NULL,
  only_feat = NULL,
  filter = 0.2,
  size = 6,
  title = NULL
)
}
\arguments{
\item{object}{An object of class 'STCompR'.}

\item{what}{A character string specifying what to plot: "changes" (the difference in pairwise Manhattan distances between conditions 2 and 1),
"changes_2" (the difference in pairwise Manhattan distances between conditions 1 and 2), "c1" (the pairwise Manhattan distances of features in condition 1), or "c2"
(the pairwise Manhattan distances of features in condition 2).}

\item{hclust_method}{A character string specifying the agglomerative criteria for hierarchical clustering ("ward.D", "ward.D2",
"single", "complete", "average", "mcquitty", "median", "centroid").}

\item{dist_method}{A character string specifying the distance method to be used ("euclidean",
"maximum", "manhattan", "canberra", "binary","minkowski").}

\item{del_feat}{A character vector specifying features to be excluded from the heatmap.}

\item{only_feat}{A character vector specifying features to be included in the heatmap.}

\item{filter}{A numeric value specifying the threshold for filtering out low values in the heatmap.}

\item{size}{A numeric value specifying the size of text in the heatmap.}
}
\value{
A heatmap plot comparing conditions in the 'STCompR' object.
}
\description{
Generate a heatmap to visualize neighborhood changes between conditions.
The changes are computed as the difference in manhattan distances between the two conditions
stored in the object.
}
\examples{
# Example usage:
example_dataset("10819270/files/cmp_xen")
heatmap_cmp(object = cmp_xen, hclust_method = "ward.D", dist_method = "euclidean")
}
\keyword{internal}
