% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{load_spatial}
\alias{load_spatial}
\title{Constructor for SpatialTranscriptomicAnalysis class (STGrid)}
\usage{
load_spatial(path = "", method = c("coordinates", "merscope"), bin_size = 25)
}
\arguments{
\item{path}{Either a file (if method is set to "coordinates") or a directory (if method
is set to "merscope"). If method is set to "coordinates" the file should contain 3 columns
("x", "y", "gene").}

\item{method}{The type of technology ("merscope" in the only supported method at the moment).}

\item{bin_size}{Numeric value representing the bin size (default to 25).}
}
\value{
An object of class STGrid.
}
\description{
Constructor for SpatialTranscriptomicAnalysis class (STGrid)
}
