% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{meta}
\alias{meta}
\title{Extract meta slot of an STGrid object}
\usage{
meta(object)
}
\arguments{
\item{object}{The STGrid object}
}
\description{
Extract meta slot of an STGrid object
}
\examples{
example_dataset()
head(meta(Xenium_Mouse_Brain_Coronal_7g))
}
\keyword{internal}
