% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{meta_names}
\alias{meta_names}
\title{The names of meta informations stored in a STGrid object}
\usage{
meta_names(object)
}
\arguments{
\item{object}{The STGrid object}
}
\description{
The names of meta informations stored in a STGrid object
}
\examples{
example_dataset()
meta_names(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
