% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{moran_index,STGrid-method}
\alias{moran_index,STGrid-method}
\title{Get Moran's index values from corresponding slot from a STGrid object.}
\usage{
\S4method{moran_index}{STGrid}(object)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Get Moran's index values from corresponding slot from a STGrid object.
}
\examples{
example_dataset()
Xenium_Mouse_Brain_Coronal_7g <- compute_k_ripley(Xenium_Mouse_Brain_Coronal_7g, verbose=FALSE)
head(moran_index(Xenium_Mouse_Brain_Coronal_7g))
}
