% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{multi_feat_contrast}
\alias{multi_feat_contrast}
\title{Multi-Gene contrast analysis for an STGrid object.}
\usage{
multi_feat_contrast(
  object,
  feat_list = NULL,
  threshold = 3,
  colors = c("black", grDevices::rainbow(15))
)
}
\arguments{
\item{object}{An STGrid object representing the spatial transcriptomics data.}

\item{feat_list}{A list of features (genes) to be analyzed. Warning needs 2^(length(feat_list)) associated colors...
A maximum of 4 genes (16 colors) is thus advised.}

\item{threshold}{A numeric value specifying the expression threshold for binarization. Default is 5 counts for a feature (e.g gene) in a x/y bin.}

\item{colors}{A vector of colors to associate with the bit code. Default is c("black", grDevices::rainbow(15)).}
}
\value{
A ggplot2 object representing the heatmap of different gene expression configurations.
}
\description{
This function performs a contrast analysis between multiple features (genes) using an STGrid as input. It binarizes the gene expression based on a user-defined threshold across x/y bins.
For each bin a binary vector is computed and encoded as a color (e.g 000, not expressing 3 user defined genes; 100, expressing gene 1, 110, expressing gene 1 and 2...).
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
multi_feat_contrast(re_bin(xen, 50), feat_list=feat_names(xen)[1:4], threshold=2)
}
\keyword{internal}
