% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{nb_feat,STCompR-method}
\alias{nb_feat,STCompR-method}
\title{The number of features stored in a STCompR object}
\usage{
\S4method{nb_feat}{STCompR}(object)
}
\arguments{
\item{object}{The STCompR object}
}
\description{
The number of features stored in a STCompR object
}
\examples{
example_dataset("11210787/files/cmp_xen")
nb_feat(cmp_xen)
}
