% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nb_feat,STGrid-method}
\alias{nb_feat,STGrid-method}
\title{The number of features stored in a STGrid object}
\usage{
\S4method{nb_feat}{STGrid}(object)
}
\arguments{
\item{object}{The STGrid object}
}
\description{
The number of features stored in a STGrid object
}
