% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R, R/st_compr_class.R
\name{nb_feat}
\alias{nb_feat}
\title{The number of features stored in a STGrid object}
\usage{
nb_feat(object)

\S4method{nb_feat}{STCompR}(object)
}
\arguments{
\item{object}{The STCompR object}
}
\description{
The number of features stored in a STGrid object

The number of features stored in a STCompR object

The number of features stored in a STCompR object
}
\keyword{internal}
