% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{nb_genes,STCompR-method}
\alias{nb_genes,STCompR-method}
\title{The number of genes stored in a STCompR object}
\usage{
\S4method{nb_genes}{STCompR}(x)
}
\arguments{
\item{x}{The STCompR object}
}
\description{
The number of genes stored in a STCompR object
}
