% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R, R/st_compr_class.R
\name{nb_genes}
\alias{nb_genes}
\title{The number of genes stored in a STGrid object}
\usage{
nb_genes(x)

nb_genes(x)

nb_genes(x)
}
\arguments{
\item{x}{The STCompR object}
}
\description{
The number of genes stored in a STGrid object

The number of genes stored in a STGrid object

The number of genes stored in a STCompR object
}
\keyword{internal}
