% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nb_items}
\alias{nb_items}
\title{The number of x/y items (e.g. molecules or cells) stored in a STGrid object}
\usage{
nb_items(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Returns the number of x/y items (e.g. molecules or cells) stored in a STGrid object.
}
\examples{
example_dataset()
nb_items(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
