% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nb_molec,STGrid-method}
\alias{nb_molec,STGrid-method}
\title{The number of molecules stored in a STGrid object}
\usage{
\S4method{nb_molec}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The number of molecules stored in a STGrid object
}
