% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nb_molec}
\alias{nb_molec}
\title{The number of molecules stored in a STGrid object}
\usage{
nb_molec(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The number of molecules stored in a STGrid object
}
\keyword{internal}
