% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nbin_x,STGrid-method}
\alias{nbin_x,STGrid-method}
\title{Number of bins (x axis) of a STGrid object.}
\usage{
\S4method{nbin_x}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Number of bins (x axis) of a STGrid object.
}
\examples{
example_dataset()
nbin_x(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
