% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nbin_y}
\alias{nbin_y}
\title{Number of bins (y axis) of a STGrid object.}
\usage{
nbin_y(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Number of bins (y axis) of a STGrid object.
}
\examples{
example_dataset()
nbin_y(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
