% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{ncol,STGrid-method}
\alias{ncol,STGrid-method}
\title{ncol}
\usage{
\S4method{ncol}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The number of columns of a STGrid object.
}
\keyword{internal}
