% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nrow,STGrid-method}
\alias{nrow,STGrid-method}
\title{nrow}
\usage{
\S4method{nrow}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The number of rows of a STGrid object.
}
\keyword{internal}
