% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{order_feat_by_moran_index,STGrid-method}
\alias{order_feat_by_moran_index,STGrid-method}
\title{Returns features ordered by Moran's Index
This method returns features ordered by Moran's Index values stored in
an \code{STGrid}.}
\usage{
\S4method{order_feat_by_moran_index}{STGrid}(object)
}
\arguments{
\item{object}{An \code{STGrid} object.}
}
\description{
Returns features ordered by Moran's Index
This method returns features ordered by Moran's Index values stored in
an \code{STGrid}.
}
