% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{order_feat_by_moran_index}
\alias{order_feat_by_moran_index}
\title{Returns features ordered by Moran's Index}
\usage{
order_feat_by_moran_index(object)
}
\arguments{
\item{object}{An \code{STGrid} object.}
}
\description{
This method returns features ordered by Moran's Index values stored in
an \code{STGrid}.
}
\keyword{internal}
