% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{order_feat_by_ripley,STGrid-method}
\alias{order_feat_by_ripley,STGrid-method}
\title{Returns features ordered by Ripley's K-function}
\usage{
\S4method{order_feat_by_ripley}{STGrid}(object)
}
\arguments{
\item{object}{An \code{STGrid} object.}
}
\description{
This method returns features ordered by Ripley's K-function results stored in
an \code{STGrid}.
}
