% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{order_feat_by_ripley}
\alias{order_feat_by_ripley}
\title{Returns features ordered by Ripley's K-function}
\usage{
order_feat_by_ripley(object)
}
\arguments{
\item{object}{An \code{STGrid} object.}
}
\description{
This method returns features ordered by Ripley's K-function results stored in
an \code{STGrid}.
}
\keyword{internal}
