% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_moran_i,STGrid-method}
\alias{plot_moran_i,STGrid-method}
\title{Plot the results from Moran's index computation stored in the STGrid object.}
\usage{
\S4method{plot_moran_i}{STGrid}(object, fill_color = "Blue", border_color = "white")
}
\arguments{
\item{object}{A STCompR object.}

\item{fill_color}{Bar color.}

\item{border_color}{Bar border color.}
}
\description{
Plot the results from Moran's index computation stored in the STGrid object.
}
