% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_rip_k,STGrid-method}
\alias{plot_rip_k,STGrid-method}
\title{Call the Ripley's k function.}
\usage{
\S4method{plot_rip_k}{STGrid}(
  object,
  correction = c("border", "isotropic", "Ripley", "translate"),
  max_feat_label = 8,
  color = NULL,
  size = 4
)
}
\arguments{
\item{object}{A STCompR object.}

\item{correction}{kind of correction should be displayed.}

\item{max_feat_label}{The maximum number of features to display.
The features with the highest correction value (whatever the radius)
are displayed.}

\item{color}{The colors for the features to be displayed.}

\item{size}{The size of the labels.}
}
\description{
Call the Ripley's k function.
}
\examples{
example_dataset()
plot_rip_k(compute_k_ripley(Xenium_Mouse_Brain_Coronal_7g[c("Ano1", "Chat", "Ebf3")]))

}
