% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_this_msg}
\alias{print_this_msg}
\title{Print a message based on the level of verbosity}
\usage{
print_this_msg(..., msg_type = c("INFO", "DEBUG", "WARNING", "STOP", "ERROR"))
}
\arguments{
\item{...}{The variables that will be pasted in the message.}

\item{msg_type}{The type of message, one of "INFO", "DEBUG",
"WARNING" or "STOP" (may also accept "ERROR" which is
translated as "STOP").}
}
\value{
None
}
\description{
Print a message based on the level of verbosity
}
\examples{
opt_warn <- options()$warn
set_verb_level(1)
print_this_msg("Hello world!", msg_type = "INFO")
set_verb_level(3)
print_this_msg("Debugging message", msg_type = "DEBUG")
set_verb_level(0)
print_this_msg("Hello world!", msg_type = "INFO")
print_this_msg("Debugging message", msg_type = "DEBUG")
options(warn=-1)
print_this_msg("A warning message not displayed", msg_type = "WARNING")
options(warn=opt_warn)
}
\keyword{internal}
