% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rand_string}
\alias{rand_string}
\title{Generate a random string of letters and numbers}
\usage{
rand_string()
}
\value{
A character string of length 10, consisting of random letters and numbers
}
\description{
This function generates a random string of 10 characters, consisting of 3 uppercase letters,
4 digits, and 3 lowercase letters.
}
\examples{
rand_string()
}
\keyword{internal}
