% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{re_bin}
\alias{re_bin}
\title{Re-bin a STGrid object.}
\usage{
re_bin(object, bin_size)
}
\arguments{
\item{object}{The STGrid object}

\item{bin_size}{The size of the bin.}
}
\description{
Re-bin a STGrid object.
}
\keyword{internal}
