% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{row_names,STGrid-method}
\alias{row_names,STGrid-method}
\title{Row names of a STGrid object.}
\usage{
\S4method{row_names}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object
example_dataset()
row_names(Xenium_Mouse_Brain_Coronal_7g)}
}
\description{
The row names of a STGrid object.
}
