% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{row_names}
\alias{row_names}
\title{Row names of a STGrid object.}
\usage{
row_names(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The row names of a STGrid object.
}
\examples{
example_dataset()
row_names(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
