% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_verb_level}
\alias{set_verb_level}
\title{Set the verbosity level for the stcompr package}
\usage{
set_verb_level(verbosity_value)
}
\arguments{
\item{verbosity_value}{A non-negative integer indicating the verbosity level to be set.}
}
\description{
This function sets the verbosity level for the stcompr package,
which controls the amount of information that is printed to the console by
the \code{\link{print_this_msg}} function. The verbosity level can be set to
any non-negative integer, with higher values indicating more detailed output.
By default, the verbosity level is set to 1.
}
\examples{
# Set verbosity level to 2
set_verb_level(2)

# Set verbosity level to 0
set_verb_level(0)
}
