% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{show,STGrid-method}
\alias{show,STGrid-method}
\title{The show method of a STGrid object}
\usage{
\S4method{show}{STGrid}(object)
}
\arguments{
\item{object}{A STGrid object.}
}
\description{
The show method of a STGrid object
}
\examples{
example_dataset()
show(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
