% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_methods}
\alias{show_methods}
\title{List the method for the STGrid object.}
\usage{
show_methods(class = "STGrid", where = "stcompr")
}
\arguments{
\item{class}{The name of the class. Default to "STGrid".}

\item{where}{The package name. Default to "STGrid".}
}
\value{
A vector with the available methods.
}
\description{
List the method for the STGrid object.
}
\examples{
show_methods()
}
