% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatial_image}
\alias{spatial_image}
\title{Create a spatial image of based on a  STGrid object.}
\usage{
spatial_image(
  object,
  gene = NULL,
  saturation = 0.75,
  scale = TRUE,
  colors = c("#000003", "#410967", "#932567", "#DC5039", "#FBA40A", "#FCFEA4"),
  ceil = TRUE,
  coord_fixed = TRUE
)
}
\arguments{
\item{object}{The STGrid object.}
}
\description{
Create a spatial image of based on a  STGrid object.
}
\keyword{internal}
