% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_gg_theming}
\alias{st_gg_theming}
\title{Custom ggplot2 Theming}
\usage{
st_gg_theming()
}
\value{
A ggplot2 theme object with predefined settings for legend, axis, and strip text.
}
\description{
This function defines a custom theme for ggplot2 plots with specific adjustments
to the legend key dimensions, legend text, axis titles, and strip text sizes.
}
\details{
The \code{st_gg_theming} function customizes the appearance of ggplot2 plots.
}
\examples{
is(st_gg_theming())

}
\keyword{internal}
