% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{stat_test,STCompR-method}
\alias{stat_test,STCompR-method}
\title{The normalized counts and test results of a STCompR object.}
\usage{
\S4method{stat_test}{STCompR}(
  object,
  transform = c("None", "log2", "log10", "log"),
  count_only = FALSE,
  melted_count = FALSE,
  normalized = TRUE,
  genes = NULL
)
}
\arguments{
\item{transform}{Whether the count are transformed (the pseudo count defined for the object is added).}

\item{count_only}{Returns only counts.}

\item{melted_count}{Returns a melted table with counts.}

\item{normalized}{Counts are normalized.}

\item{genes}{The genes that should be returned. Default to all (NULL).}

\item{x}{The STCompR object}
}
\description{
The normalized counts and test results of a STCompR object.
}
