% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{stat_test}
\alias{stat_test}
\title{The normalized counts and test results of a STCompR object.}
\usage{
stat_test(x)

stat_test(x)
}
\arguments{
\item{x}{The STCompR object}
}
\description{
The normalized counts and test results of a STCompR object.

The normalized counts and test results of a STCompR object.
}
\keyword{internal}
