% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{stgrid_from_data_frame}
\alias{stgrid_from_data_frame}
\title{Create an STGrid from a data.frame}
\usage{
stgrid_from_data_frame(this_df = NULL, mapping = NULL, bin_size = 25)
}
\arguments{
\item{this_df}{A data.frame}

\item{mapping}{Please provide a named vector indicating which column map to "x", "y", and "feature" in the STGrid object.}
}
\value{
An STGrid object.
}
\description{
Create an STGrid from a data.frame
}
\examples{
d <- data.frame(foo=runif(1000, 0, 1000), bar=runif(1000, 0, 1000), bla=sample(letters, 1000, replace=TRUE))
st <- stgrid_from_data_frame(d, mapping=c("x"="foo", "y"="bar", "feature"="bla"))
st
}
