% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{[,STGrid,ANY,ANY,ANY-method}
\alias{[,STGrid,ANY,ANY,ANY-method}
\title{Subsetting operator of a STGrid object
Extract}
\usage{
\S4method{[}{STGrid,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{...}{See ?'['. Not functionnal here.}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension. Not functionnal here.}

\item{features}{(i.e. genes) or bins (x axis) to extract or substract.}
}
\description{
The subsetting operator of a STGrid object.
The i axis correspond to features (i.e. genes) or bins (x axis). The j axis corresponds to bins
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
xen["Ano1",]
xen[bin_x(xen)[1:20], bin_y(xen)[1:20]]
}
\keyword{internal}
