% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{[,STGrid-method}
\alias{[,STGrid-method}
\title{Subsetting operator of a STGrid object
Extract}
\usage{
\S4method{[}{STGrid}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{i}{indices specifying genes to extract or substract.}

\item{...}{See ?'['. Not functionnal here.}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension. Not functionnal here.}
}
\description{
The subsetting operator of a STGrid object.
The i axis correspond to genes. j axis is not implemented yet.
}
\keyword{internal}
