% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{[[<-,STGrid,character,ANY,ANY-method}
\alias{[[<-,STGrid,character,ANY,ANY-method}
\title{Replacing operator "[[" of a STGrid object. Extract metadata.}
\usage{
\S4method{[[}{STGrid,character,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{value}{The replacement value.}

\item{name}{The name of the metadata.}
}
\description{
Subsetting operator "[[" of a STGrid object. Extract metadata.
}
\examples{
example_dataset()
Xenium_Mouse_Brain_Coronal_7g[[1]] <- Xenium_Mouse_Brain_Coronal_7g[[1]]
}
\keyword{internal}
