% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{[[<-,STGrid-method}
\alias{[[<-,STGrid-method}
\title{Replacing operator "[[" of a STGrid object. Extract metadata.}
\usage{
\S4method{[[}{STGrid}(x, name) <- value
}
\arguments{
\item{name}{The name of the metadata.}

\item{value}{The replacement value.}
}
\description{
Subsetting operator "[[" of a STGrid object. Extract metadata.
}
\keyword{internal}
