% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{sum_of_counts}
\alias{sum_of_counts}
\title{Compute the Sum of Counts}
\usage{
sum_of_counts(spatial_matrix = NULL, control = NULL)
}
\arguments{
\item{spatial_matrix}{The spatial matrix containing molecule counts, typically obtained from a spatial transcriptomics experiment.}

\item{control}{A regular expression pattern indicating the controls in the dataset. Controls are internal features used for for monitoring experiments.}
}
\value{
A numeric vector containing the sum of counts for each row in the spatial matrix.
}
\description{
This function calculates the sum of counts for each row in a spatial matrix.
}
