% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{summary,STGrid-method}
\alias{summary,STGrid-method}
\title{The summary() method of a STGrid object}
\usage{
\S4method{summary}{STGrid}(object)
}
\arguments{
\item{object}{A STGrid object.}
}
\description{
The summary method of a STGrid object
}
