% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{tab,STGrid-method}
\alias{tab,STGrid-method}
\title{The number of items (molecules/cells) per features (gene/cell-type).}
\usage{
\S4method{tab}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Returns The number of items (molecules/cells) per features (gene/cell-type).
}
\examples{
example_dataset()
tab(Xenium_Mouse_Brain_Coronal_7g)
}
