% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{table_st}
\alias{table_st}
\title{The number of items (e.g molecules) per feature of an STGrid object.}
\usage{
table_st(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Returns the number of items (e.g molecules) per feature of an STGrid object.
}
\examples{
example_dataset()
table(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
