% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{write_coord}
\alias{write_coord}
\title{Write object coordinates.}
\usage{
write_coord(object = NULL, file_path = NULL)
}
\arguments{
\item{object}{The STGrid object}

\item{file_path}{The path where to store the file.}
}
\description{
Write object coordinates.
}
\examples{
example_dataset()
tmp_f <- tempfile()
write_coord(Xenium_Mouse_Brain_Coronal_7g, file_path=tmp_f)
}
\keyword{internal}
