% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{bin_mat}
\alias{bin_mat}
\title{The binned matrix stored in a STGrid object}
\usage{
bin_mat(
  object,
  as_factor = FALSE,
  feat_list = character(),
  melt_tab = FALSE,
  del_bin = FALSE,
  pseudo_count = 0,
  transform = c("None", "log2", "log10", "log"),
  axis_as_number = FALSE
)
}
\arguments{
\item{object}{The STGrid object}

\item{as_factor}{Should bin_x and bin_y columns be returned as ordered factor ?}

\item{feat_list}{Whether to subset to some features.}

\item{melt_tab}{Whether to melt.}

\item{del_bin}{Whether to delete the bin_x, bin_y columns.}

\item{pseudo_count}{Whether to add a pseudo_count.}

\item{transform}{Whether the count should be transformed.}

\item{axis_as_number}{If melt_tab is TRUE, whether bin_x and bin_y should be transformed to numerics.}
}
\description{
The binned matrix stored in a STGrid object
}
\examples{
example_dataset()
head(bin_mat(Xenium_Mouse_Brain_Coronal_7g))
}
\keyword{internal}
