% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{cc_neighborhood}
\alias{cc_neighborhood}
\title{Find the neighbors of a Connected Component Using Queen Criterion}
\usage{
cc_neighborhood(object, feature = NULL, neighborhood_size = 0:4)
}
\arguments{
\item{object}{An STGrid object containing the spatial expression data.}

\item{feature}{A character string specifying the feature of interest.}

\item{neighborhood_size}{An integer or a vector of integers specifying the neighborhood size(s) to consider. Default is 1:4.}
}
\value{
The input STGrid object with additional fields (in meta slot) indicating the satellite bins for the specified feature at each neighborhood size.
}
\description{
Given an STGrid object and a Connected Component (see connected_components()) find the neighboring bins.
}
\examples{
library(patchwork)
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
p1 <- spatial_image(xen, feat="Necab2")
xen <- connected_components(xen, feat_list = "Necab2", threshold = 5, min_size = 30)
head(meta(xen))
p2 <- spatial_image(xen, feat="Necab2_cpt", as_factor=TRUE, colors = c("black", rainbow(9)))
hull_1 <- create_hull(xen, feat="Necab2_cpt", color="red", linew=0.3)
p3 <- p2 + hull_1
p4 <- spatial_image(xen, feat="Nwd2") + hull_1
xen <- cc_neighborhood(xen, feat="Necab2_cpt", neighborhood_size=1:3)
head(meta(xen))
hull_2 <- create_hull(xen, feat="Necab2_cpt_ns_1", color="green", linew=0.3)
p5 <- spatial_image(xen, feat="Necab2_cpt", as_factor=TRUE, colors = c("black", rainbow(9))) + hull_2
hull_3 <- create_hull(xen, feat="Necab2_cpt_ns_3", color="green", linew=0.3)
p6 <- spatial_image(xen, feat="Necab2_cpt", as_factor=TRUE, colors = c("black", rainbow(9))) + hull_3
(p1 | p2 ) / (p3 | p4) / (p5 + p6)
}
