% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_bar_plot}
\alias{cmp_bar_plot}
\title{Create a barplot to show counts for selected features.}
\usage{
cmp_bar_plot(
  object,
  features = utils::head(feat_names(object)),
  normalized = FALSE,
  transform = c("None", "log2", "log10", "log"),
  colors = c("#3074BB", "#BE5B52")
)
}
\arguments{
\item{object}{A STGrid object.}

\item{features}{The list of features (NULL for all of them).}

\item{normalized}{Whether counts should be normalized.}

\item{transform}{Whether the count should be transformed (the pseudo count defined for the object is added).}

\item{colors}{A set of colors.}
}
\description{
Create a barplot to show counts for selected features.
}
\examples{
example_dataset()
xen <- Xenium_Mouse_Brain_Coronal_7g
x_bins <-  bin_x(xen)[181:nbin_x(xen)]
y_bins <-  bin_y(xen)[101:nbin_y(xen)]
xen_r1 <- xen[x_bins, y_bins]
x_bins <-  bin_x(xen)[61:101]
y_bins <-  bin_y(xen)[101:nbin_y(xen)]
xen_r2 <- xen[x_bins, y_bins]
cmp <- stcompr(xen_r1, xen_r2)
cmp_bar_plot(cmp,
             features=c("Chat", "Nwd2", "Ano1"))
}
\keyword{internal}
