% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cmp_volcano}
\alias{cmp_volcano}
\title{Create a volcano plot to compare molecule counts between 2 conditions.}
\usage{
cmp_volcano(
  object,
  x_axis = c("log2_ratio", "log2_odds"),
  y_axis = c("p_values", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"),
  x_lim = c(-2.5, 2.5),
  colors = RColorBrewer::brewer.pal(7, "Spectral"),
  text_y_lim = 100,
  text_x_lim = 2,
  text_size = 5,
  title = NULL,
  max.overlaps = 10
)
}
\arguments{
\item{object}{A STCompR object.}

\item{x_axis}{Character vector specifying the x-axis variable.
Options are "log2_ratio" or "log2_odds".}

\item{y_axis}{Character vector specifying the y-axis variable.
Options are "p_values", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", or "fdr".}

\item{x_lim}{Numeric vector specifying the x-axis limits. Default is c(-2.5, 2.5).}

\item{colors}{Color palette for the plot. Default is RColorBrewer::brewer.pal(7, "Spectral").}

\item{text_y_lim}{Numeric value specifying the threshold for feature labels on the y-axis. Features with -log10(p-value) less than this threshold won't be labeled. Default is 100.}

\item{text_x_lim}{Numeric value specifying the threshold for feature labels on the x-axis. Features with absolute x-values less than this threshold won't not be labeled. Default is 2.}

\item{text_size}{Numeric value specifying the size of text labels in the plot. Default is 5.}

\item{title}{A title for the diagram.}

\item{max.overlaps}{The maximum number of label overlaps.}
}
\description{
Create a boxplot/jitter plot to show molecule counts distribution.
}
\examples{
example_dataset("10819270/files/cmp_xen")
cmp_xen
cmp_volcano(cmp_xen)

}
\keyword{internal}
