% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{coord}
\alias{coord}
\title{The coordinates stored in a STGrid object}
\usage{
coord(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The coordinates stored in a STGrid object
}
\keyword{internal}
