% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{dim,STGrid-method}
\alias{dim,STGrid-method}
\title{Dimension of a STGrid object.
dim}
\usage{
\S4method{dim}{STGrid}(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
The dimension of a STGrid object.
}
\keyword{internal}
