% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R, R/st_compr_class.R
\name{gene_names}
\alias{gene_names}
\title{The genes stored in a STGrid object}
\usage{
gene_names(x)

gene_names(x)

gene_names(x)
}
\arguments{
\item{x}{The STCompR object}
}
\description{
The genes stored in a STGrid object

The genes stored in a STGrid object

The genes stored in a STCompR object
}
\keyword{internal}
