% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{hc_tree}
\alias{hc_tree}
\title{Create a tree from an STGrid object}
\usage{
hc_tree(
  object,
  method = c("complete", "ward.D", "ward.D2", "single", "average", "mcquitty", "median",
    "centroid"),
  layout = c("circular", "rectangular", "slanted", "fan", "unrooted", "time-scale",
    "two-dimensional"),
  dist_method = "pearson",
  branch_length = "none",
  class_nb = 1,
  class_name = NULL,
  size = 2.25,
  colors = NULL,
  class_label = TRUE,
  offset = 5,
  offset.text = 1.5,
  lab_fontsize = 3.88,
  lab_barsize = 2,
  geom_label = c("text", "label", "shadowtext"),
  no_legend = FALSE,
  return_list = FALSE
)
}
\arguments{
\item{object}{The STGrid object}

\item{method}{The agglomeration method to be used. See stats::hclust().}

\item{layout}{The layout to be used See ggtree::ggtree().}

\item{dist_method}{The method for distance computation. See stats::cor().}

\item{branch_length}{A variable for scaling branch, if 'none' draw cladogram. See ggtree::ggtree().}

\item{class_nb}{An integer indicating the desired number of groups.}

\item{class_name}{A set of names for the classes.}

\item{size}{The size of the labels.}

\item{colors}{A set of colors for the classes.}

\item{class_label}{Whether to add class labels.}

\item{offset}{Offset of labels from the clade}

\item{offset.text}{Offset of text labels from the bar.}

\item{lab_fontsize}{Label font size.}

\item{lab_barsize}{Bar label size.}

\item{geom_label}{Label fomat.}

\item{no_legend}{Whether to discard legend.}

\item{return_list}{Logical. Whether to return a list with the tree,
associated clusters, distance and correlation matrices.}
}
\description{
Create a tree from an STGrid object
}
\examples{
example_dataset()
p <- hc_tree(Xenium_Mouse_Brain_Coronal_7g)
print(p)
}
\keyword{internal}
