% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{load_spatial}
\alias{load_spatial}
\title{Create a Spatial Transcriptomic Grid class (STGrid)}
\usage{
load_spatial(
  path = "",
  method = c("coordinates", "merscope", "merscope_csv", "xenium", "cosmx"),
  bin_size = 25,
  control = NULL,
  sep = "\\t",
  threads = 1,
  constrain = NULL,
  verbose = TRUE,
  mapping = NULL
)
}
\arguments{
\item{path}{Either a file (if method is set to "coordinates") or a directory (if method
is set to "merscope"). If method is set to "coordinates" the file should contain at least 3 columns
("x", "y", "gene"), ("x", "y", "feature") or ("x", "y", "cell").}

\item{method}{The type of technology/file/directory structure. See details.}

\item{bin_size}{Numeric value representing the bin size (default to 25). Unit is microns for Xenium and Merscope and pixel for CosMx.}

\item{control}{A regular expression to identify controls. As the function computes the sum of
counts, this will allow to delete these blanks/controls for computation.}

\item{sep}{The separator when method is set to "coordinates" (default "\\t").}

\item{threads}{The number of threads (see data.table::fread).}

\item{constrain}{Whether to put constrains on a column when method is set to 'coordinates'. E.g: 'list("global_z==0", "fov==0")'.}

\item{verbose}{Whether to display the progress bar.}

\item{mapping}{if method='coordinates' is used and non conventional column names are used in the input file,}
}
\value{
An object of class STGrid.
}
\description{
The load_spatial() function is the entry point of the STarlight package.
It will load the molecule coordinates and create a 2D binned grid with default size 25µm.
}
\details{
If method is set to 'coordinates' a flat file with ("x", "y", "gene"), ("x", "y", "feature") or ("x", "y", "cell")
expected for path. If method is set to 'merscope_csv' the transcript csv file exported by Merscope should be provided
as path (contains 'global_x', 'global_y' and 'gene' column). If method is set to 'xenium' or 'merscope'
path is passed to the 'data.dir' argument of Seurat::ReadXenium and Seurat::ReadVizgen respectively (see corresponding docs).
}
\examples{
  fp <- file.path(system.file("extdata", package = "STarlight"), "tyni_xenium.txt")
  st <- load_spatial(fp, method = "coordinates")

}
