% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{load_spatial}
\alias{load_spatial}
\title{Create a Spatial Transcriptomic Grid class (STGrid)}
\usage{
load_spatial(
  path = "",
  method = c("coordinates", "merscope", "xenium"),
  bin_size = 25,
  control = NULL,
  sep = "\\t",
  threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{path}{Either a file (if method is set to "coordinates") or a directory (if method
is set to "merscope"). If method is set to "coordinates" the file should contain at least 3 columns
("x", "y", "gene"), ("x", "y", "feature") or ("x", "y", "cell"). However, when column contains "global_x",
"global_y" and "gene", "global_x" and "global_y" will be prefered to "x" and "y" (to support Merscope csv format).}

\item{method}{The type of technology.}

\item{bin_size}{Numeric value representing the bin size (default to 25).}

\item{control}{A regular expression to identify controls. As the function computes the sum of
counts, this will allow to delete these blanks/controls for computation.}

\item{sep}{The separator when method is set to "coordinates" (default "\\t").}

\item{threads}{The number of threads (see data.table::fread).}

\item{verbose}{Whether to display the progress bar.}
}
\value{
An object of class STGrid.
}
\description{
The load_spatial() function is the entry point of the stcompr package.
It will load the molecule coordinates and create a 2D binned grid with default size 25µm.
}
\examples{
  fp <- file.path(system.file("extdata", package = "stcompr"), "tyni_xenium.txt")
  st <- load_spatial(fp, method = "coordinates")

}
