% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nb_feat}
\alias{nb_feat}
\title{The number of features stored in a STGrid object}
\usage{
nb_feat(object)
}
\arguments{
\item{object}{The STGrid object}
}
\description{
The number of features stored in a STGrid object
}
\examples{
example_dataset()
nb_feat(Xenium_Mouse_Brain_Coronal_7g)
}
\keyword{internal}
