% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{nbin_x}
\alias{nbin_x}
\title{Number of bins (x axis) of a STGrid object.}
\usage{
nbin_x(x)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Number of bins (x axis) of a STGrid object.
}
\keyword{internal}
