% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_rip_k,STGrid-method}
\alias{plot_rip_k,STGrid-method}
\title{Call the Ripley's k function.}
\usage{
\S4method{plot_rip_k}{STGrid}(
  object,
  correction = c("border", "isotropic", "Ripley", "translate"),
  max_gene_label = 8,
  color = NULL
)
}
\arguments{
\item{object}{A STCompR object.}

\item{max_gene_label}{The maximum number of genes to display.
The genes with the highest correction value (whatever the radius)
are displayed.}

\item{color}{The colors for the genes to be displayed.}

\item{Which}{kind of correction should be displayed.}
}
\description{
Call the Ripley's k function.
}
