% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_rip_k}
\alias{plot_rip_k}
\title{Plot the results from the Ripley's k function stored in the STGrid object.}
\usage{
plot_rip_k(
  object,
  correction = c("border", "isotropic", "Ripley", "translate"),
  max_gene_label = 8,
  color = NULL,
  size = 4
)
}
\arguments{
\item{object}{A STCompR object.}

\item{max_gene_label}{The maximum number of genes to display.
The genes with the highest correction value (whatever the radius)
are displayed}

\item{color}{The colors for the genes to be displayed.}

\item{size}{The size of the labels.}

\item{Which}{kind of correction should be displayed.}
}
\description{
Plot the results from the Ripley's k function stored in the STGrid object.
}
\keyword{internal}
