% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{ripley_k_function,STGrid-method}
\alias{ripley_k_function,STGrid-method}
\title{Get Ripley's K function slot from a STGrid object.}
\usage{
\S4method{ripley_k_function}{STGrid}(object)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Get Ripley's K function slot from a STGrid object.
}
\examples{
example_dataset()
Xenium_Mouse_Brain_Coronal_7g <- compute_k_ripley(Xenium_Mouse_Brain_Coronal_7g)
head(ripley_k_function(Xenium_Mouse_Brain_Coronal_7g))
}
\keyword{internal}
