% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{ripley_k_function}
\alias{ripley_k_function}
\title{Get Ripley's K function slot from a STGrid object.}
\usage{
ripley_k_function(object)
}
\arguments{
\item{x}{The STGrid object}
}
\description{
Get Ripley's K function slot from a STGrid object.
}
\keyword{internal}
