% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{rm_controls,STGrid-method}
\alias{rm_controls,STGrid-method}
\title{Remove control (Blank-*) "genes" from a STGrid object.}
\usage{
\S4method{rm_controls}{STGrid}(object, regexp = NULL)
}
\arguments{
\item{object}{The STGrid object}
}
\description{
Remove control (Blank-*) "genes".
}
