% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{rm_controls}
\alias{rm_controls}
\title{Remove control (Blank-*) "genes" from a STGrid object.}
\usage{
rm_controls(object, regexp = NULL)
}
\arguments{
\item{object}{The STGrid object}
}
\description{
Remove control (Blank-*) "genes".
}
\keyword{internal}
