% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{rm_controls}
\alias{rm_controls}
\title{Remove control (Blank-*) features from a STGrid object.}
\usage{
rm_controls(object, regexp = NULL)
}
\arguments{
\item{object}{The STGrid object}

\item{regexp}{A regexp to search for internal controls.}
}
\description{
Remove control (Blank-*) features.
}
\keyword{internal}
