% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{show,STCompR-method}
\alias{show,STCompR-method}
\title{The show method of a STCompR object}
\usage{
\S4method{show}{STCompR}(object)
}
\arguments{
\item{object}{A STCompR object.}
}
\description{
The show method of a STCompR object displays information about the object,
including conditions, memory usage, and the number of genes.
}
\examples{
# Example usage:
show(st_compr_result)

}
\keyword{internal}
