% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatial_plot}
\alias{spatial_plot}
\title{Plot x/y coordinates of molecules}
\usage{
spatial_plot(
  object = NULL,
  gene_list = NULL,
  colors = NULL,
  size = 0.1,
  coord_fixed = TRUE
)
}
\arguments{
\item{object}{The STGrid object.}

\item{colors}{The colors to use for genes in the spatial image.}

\item{size}{The size of the points}

\item{coord_fixed}{Logical value indicating whether to keep the aspect ratio fixed. Defaults to TRUE.}

\item{gene}{The name of the genes for which the spatial image will be created.}
}
\description{
Plot x/y coordinates of molecules of a spatial transcriptomics experiment.
}
\keyword{internal}
