% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{stat_test}
\alias{stat_test}
\title{The normalized counts and test results of a STCompR object.}
\usage{
stat_test(
  object,
  transform = c("None", "log2", "log10", "log"),
  count_only = FALSE,
  melted_count = FALSE,
  normalized = TRUE,
  features = NULL
)
}
\arguments{
\item{object}{The STCompR object}

\item{transform}{Whether the count are transformed (the pseudo count defined for the object is added).}

\item{count_only}{Returns only counts.}

\item{melted_count}{Returns a melted table with counts.}

\item{normalized}{Counts are normalized.}

\item{features}{The features that should be returned. Default to all (NULL).}
}
\description{
The normalized counts and test results of a STCompR object.
}
\examples{
example_dataset("11210787/files/cmp_xen")
stat_test(cmp_xen)
}
\keyword{internal}
