% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\docType{class}
\name{STCompR-class}
\alias{STCompR-class}
\title{Spatial transcriptomic comparison class (STCompR).}
\description{
This class represents the result of a spatial transcriptomic comparison between
two STGrid objects. It stores information such as conditions, methods, raw counts,
scaling factors, and statistical test results.
}
\section{Slots}{

\describe{
\item{\code{conditions}}{A character vector specifying the conditions being compared.}

\item{\code{mode}}{A character vector specifying the mode used for loading (see load_spatial()).}

\item{\code{raw_counts}}{A data.frame storing raw count data for each condition.}

\item{\code{norm_counts}}{A data.frame storing normalized counts for each condition.}

\item{\code{scaling_factor}}{A numeric vector representing the scaling factors used during normalization (to be divided).}

\item{\code{stat_test}}{A data frame containing statistical test results, including log2 ratios,
odds ratios, and p-values for each feature.}

\item{\code{pseudo_count}}{The value for the pseudo_count.}

\item{\code{neighborhood}}{a list of feature-feature neighborhood for the two conditions to be compared (not functional at the moment).}

\item{\code{neighborhood_changes}}{A feature-feature matrix indicative of neighborhood changes (not functional at the moment).}
}}

\examples{
# Example usage:
example_dataset("11284296/files/cmp_xen")
cmp_xen

}
