% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compr_class.R
\name{stcompr}
\alias{stcompr}
\title{Create a STCompR compare class to compare two STGrid objects.}
\usage{
stcompr(
  object_1,
  object_2,
  name_1 = "Condition_1",
  name_2 = "Condition_2",
  stat_test = c("Fisher"),
  norm_method = c("RLE"),
  pseudo_count = 1
)
}
\arguments{
\item{object_1}{A STGrid object.}

\item{object_2}{A STGrid object.}

\item{name_1}{A character string specifying the name for Condition 1.}

\item{name_2}{A character string specifying the name for Condition 2.}

\item{stat_test}{A character string specifying the statistical test to be used.
Currently, only "Fisher" is supported.}

\item{norm_method}{A character string specifying the normalization method to be used.
Currently, only "RLE" is supported.}

\item{pseudo_count}{A numeric value specifying the pseudo-count to be added during normalization.}
}
\value{
An object of class 'STCompR'.
}
\description{
Create a STCompR compare class to compare two STGrid objects.
}
\examples{
# Example usage:
st_grid_1 <- load_spatial(...)
st_grid_2 <- load_spatial(...)
st_compr_result <- stcompr(st_grid_1, st_grid_2, name_1 = "Condition1", name_2 = "Condition2")
}
