% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\docType{class}
\name{STGrid-class}
\alias{STGrid-class}
\title{Spatial transcriptomic anlysis class (STGrid).}
\description{
A class for representing spatial transcriptomic analysis results using (currently) the "merscope" method.
This class store transcript molecule coordinates as a grid.
}
\section{Slots}{

\describe{
\item{\code{coordinate}}{A list of data frames representing spatial coordinates (empty by default).
May accept several layers in the future.}

\item{\code{bin_mat}}{A list of data frames representing binned matrices (empty by default).
May accept several layers in the future.}

\item{\code{y_max}}{Numeric value representing the maximum y-coordinate (default is 0).}

\item{\code{y_min}}{Numeric value representing the minimum y-coordinate (default is 0).}

\item{\code{x_min}}{Numeric value representing the minimum x-coordinate (default is 0).}

\item{\code{x_max}}{Numeric value representing the maximum x-coordinate (default is 0).}

\item{\code{path}}{Character value representing the file path to the original data.}

\item{\code{meta}}{A list containing meta-information.}

\item{\code{bin_size}}{Numeric value representing the bin size.}

\item{\code{bin_x}}{Character vector containing the names of bins/windows along the x-axis.}

\item{\code{bin_y}}{Character vector containing the names of bins/windows along the y-axis.}
}}

