% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\docType{class}
\name{STGrid-class}
\alias{STGrid-class}
\title{Spatial transcriptomic analysis class (STGrid).}
\description{
A class for representing spatial transcriptomic analysis results as x/y coordinates.
This class stores transcript molecules or cell coordinates as a grid.
}
\section{Slots}{

\describe{
\item{\code{coord}}{A data frames representing spatial coordinates (empty by default).}

\item{\code{bin_mat}}{A data frames representing the matrix (empty by default).}

\item{\code{y_max}}{Numeric value representing the maximum y-coordinate (default is 0).}

\item{\code{y_min}}{Numeric value representing the minimum y-coordinate (default is 0).}

\item{\code{x_min}}{Numeric value representing the minimum x-coordinate (default is 0).}

\item{\code{x_max}}{Numeric value representing the maximum x-coordinate (default is 0).}

\item{\code{path}}{Character value representing the file path to the original data.}

\item{\code{method}}{The technology used.}

\item{\code{meta}}{A list containing meta-information about bins.}

\item{\code{bin_size}}{Numeric value representing the bin size.}

\item{\code{bin_x}}{Character vector containing the names of bins/windows along the x-axis.}

\item{\code{bin_y}}{Character vector containing the names of bins/windows along the y-axis.}

\item{\code{ripley_k_function}}{The result of Ripley's K function computation.}

\item{\code{control}}{A regular expression indicating the string motif related to controls.}
}}

\examples{
example_dataset()
Xenium_Mouse_Brain_Coronal_7g
show_methods(class = "STGrid")
}
