% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_class.R
\name{[[,STGrid,ANY,ANY-method}
\alias{[[,STGrid,ANY,ANY-method}
\title{Subsetting operator "[[" of a STGrid object. Extract metadata.}
\usage{
\S4method{[[}{STGrid,ANY,ANY}(x, i, exact = TRUE)
}
\arguments{
\item{i}{indices specifying metadata to extract.}

\item{exact}{See ?"[[".}
}
\description{
Subsetting operator "[[" of a STGrid object. Extract metadata.
}
\keyword{internal}
